requirements.txt

argparse
numpy
opencv-python
torch
tqdm
scipy
matplotlib
pyyaml
scikit-learn
